Update MANIFESTO.md; Rename lists to match JVM
