+++ src/zepben/ewb/model/cim/extensions/iec61968/assetinfo/relay_info.py
@@ -13,7 +13,7 @@ from typing import Optional, List, Callable, Any
 from typing_extensions import deprecated
 
 from zepben.ewb.boilerplate.dataclass_base import zb_dataclass
-from zepben.ewb.boilerplate.collections.lazy_list import LazyList
+from zepben.ewb.boilerplate.collections.lazy_index_list import LazyIndexList
 from zepben.ewb.model.cim.extensions.zbex import zbex
 from zepben.ewb.model.cim.iec61968.assets.asset_info import AssetInfo
 
@@ -31,7 +31,7 @@ class RelayInfo(AssetInfo):
 
     _reclose_delays: Optional[List[float]] = field(default=None)
 
-    reclose_delays: LazyList[float] = LazyList(
+    reclose_delays: LazyIndexList[float] = LazyIndexList(
         _reclose_delays,
         "float"
     )
