+++ src/zepben/ewb/model/cim/iec61970/base/core/connectivity_node.py
@@ -14,6 +14,7 @@ from typing_extensions import deprecated
 
 from zepben.ewb.boilerplate.collections.lazy_mrid_list import LazyMridList
 from zepben.ewb.boilerplate.collections.mrid_collection import MridCollection
+from zepben.ewb.boilerplate.collections.mrid_list import MridList
 from zepben.ewb.model.cim.iec61970.base.core.identified_object import IdentifiedObject
 from zepben.ewb.boilerplate.dataclass_base import zb_dataclass
 
@@ -36,7 +37,7 @@ class ConnectivityNode(IdentifiedObject, WeakrefSlot):
     def __iter__(self):
         return iter(self._terminals)
 
-    terminals: MridCollection[Terminal] = LazyMridList(
+    terminals: MridCollection[Terminal] = MridList(
         _terminals,
         "A Terminal"
     )
